% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areas_update.R
\name{areas.sum_to}
\alias{areas.sum_to}
\title{Areas summing up function over land-use changes}
\usage{
areas.sum_to(res, curr.areas, priors, xmat, xmat.proj)
}
\arguments{
\item{res}{Result from downscale}

\item{curr.areas}{Dataframe of current areas}

\item{priors}{Priors for updating}

\item{xmat}{X-matrix dataframe}

\item{xmat.proj}{Projected x matrix}
}
\value{
curr.areas A
}
\description{
Areas summing up function over land-use changes
}
