% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{argentina_df}
\alias{argentina_df}
\title{Argentina land-use change driver data}
\format{
A list with four items: \code{xmat}, \code{restrictions}, \code{pop_data}, \code{lu_levels}
\code{xmat} is a  data frame with 96400 rows and 3 variables:
\describe{
\item{ns}{Pixel ids for Argentina}
\item{ks}{Variable names}
\item{value}{Value of variable}
}
\code{restrictions} is a  data frame with 38560 rows and 4 variables:
\describe{
\item{ns}{Pixel ids for Argentina}
\item{lu.from}{Origin land-use class of converion}
\item{lu.to}{Destination land-use class to which lu.from is converted}
\item{value}{Binary, 1 denotes that the conversion lu.from to lu.to is not allowed}
}
\code{pop_data} is a  data frame with 84832 rows and 4 variables, containing population projections until 2050
\describe{
\item{ns}{Pixel ids for Argentina}
\item{times}{Year of population projection}
\item{ks}{totPop - total population or ruralPop - rural population}
\item{value}{Number of persons}
}
\code{lu_levels} is a  data frame with 23136 rows and 3 variables, containing land-use observations in 2000
\describe{
\item{ns}{Pixel ids for Argentina}
\item{lu.from}{Land-use class}
\item{value}{Ha of land}
}
}
\description{
Argentina land-use change driver data
}
\keyword{2000}
\keyword{Argentina}
\keyword{Data}
\keyword{argentina}
\keyword{as}
\keyword{change}
\keyword{drivers,}
\keyword{in}
\keyword{land}
\keyword{land-use}
\keyword{land-uses}
\keyword{observed}
\keyword{of}
\keyword{use}
\keyword{well}
