% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{argentina_FABLE}
\alias{argentina_FABLE}
\title{FABLE Calculator projections for Argentina}
\format{
A data frame with 150 rows and 4 variables:
\describe{
\item{times}{Time period of land-use changes}
\item{lu.from}{Origin land-use class of converion}
\item{lu.to}{Destination land-use class to which lu.from is converted}
\item{value}{Ha land being converted from lu.from to lu.to in time period}
}
}
\source{
\url{https://www.abstract-landscapes.com/fable-calculator}
}
\description{
FABLE Calculator projections for Argentina
}
\keyword{2050}
\keyword{Calculator}
\keyword{FABLE}
\keyword{Land-use}
\keyword{change}
\keyword{from}
\keyword{land}
\keyword{projections}
\keyword{the}
\keyword{until}
\keyword{use}
