% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{argentina_luc}
\alias{argentina_luc}
\title{Land-use change in Argentina from 2000 to 2020}
\format{
A data frame with 277632 rows and 5 variables:
\describe{
\item{ns}{Pixel ids for Argentina}
\item{lu.from}{Origin land-use class of converion}
\item{lu.to}{Destination land-use class to which lu.from is converted}
\item{Ts}{Time-period; either 2000 (changes 2000 to 2010) or 2010 (changes 2010 to 2020)}
\item{value}{Percent of lu.from in pixel being converted}
}
}
\source{
MapBiomas
}
\description{
Land-use change in Argentina from 2000 to 2020
}
\keyword{2000}
\keyword{2010}
\keyword{2010,}
\keyword{2020.}
\keyword{Argentina}
\keyword{Data}
\keyword{and}
\keyword{as}
\keyword{between}
\keyword{change}
\keyword{changes}
\keyword{in}
\keyword{land-use}
\keyword{landuse}
\keyword{of}
\keyword{well}
