% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{argentina_raster}
\alias{argentina_raster}
\title{List of Argentina key raster objects}
\format{
A list with 3 objects:
\describe{
\item{CRS}{Coordinate Reference System object}
\item{values}{A matrix with 431 rows and 237 cols, !NA elements grid location of Argentina}
\item{extent}{Extent object for Argentina raster map}
}
}
\source{
GLOBIOM map
}
\description{
List of Argentina key raster objects
}
\keyword{Argentina}
\keyword{CRS}
\keyword{List}
\keyword{argentina}
\keyword{extent}
\keyword{map.}
\keyword{object}
\keyword{object,}
\keyword{of}
\keyword{raster}
\keyword{reconstruct}
\keyword{to}
\keyword{values,}
