% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_input.R
\name{complete_areas}
\alias{complete_areas}
\title{Complete input areas}
\usage{
complete_areas(areas)
}
\arguments{
\item{areas}{Dataframe of areas, must have columns ns and value}
}
\value{
Completed dataframe with lu.from and all combinations

Internal function. Adds missing columns and completes potential sparse dataframes.
}
\description{
Complete input areas
}
\keyword{internal}
