% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_input.R
\name{complete_betas}
\alias{complete_betas}
\title{Complete input betas}
\usage{
complete_betas(betas)
}
\arguments{
\item{betas}{Dataframe of betas, must have columns ks, lu.from and lu.to and value}
}
\value{
Completed dataframe, with added lu.from

Internal function. Adds missing columns and completes potential sparse dataframes.
}
\description{
Complete input betas
}
\keyword{internal}
