% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_input.R
\name{complete_priors}
\alias{complete_priors}
\title{Complete input priors}
\usage{
complete_priors(priors)
}
\arguments{
\item{priors}{Dataframe of priors, must have columns ns, lu.to and value}
}
\value{
Completed dataframe with lu.from and all combinations

Internal function. Adds missing columns and completes potential sparse dataframes.
}
\description{
Complete input priors
}
\keyword{internal}
