% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_input.R
\name{complete_restrictions}
\alias{complete_restrictions}
\title{Complete input restrictions}
\usage{
complete_restrictions(restrictions)
}
\arguments{
\item{restrictions}{Dataframe of restrictions, must have columns ns, lu.to and value}
}
\value{
Completed dataframe with lu.from and all combinations

Internal function. Adds missing columns and completes potential sparse dataframes.
}
\description{
Complete input restrictions
}
\keyword{internal}
