% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_input.R
\name{complete_targets}
\alias{complete_targets}
\title{Complete input targets}
\usage{
complete_targets(targets)
}
\arguments{
\item{targets}{Dataframe of targets, must have columns lu.to and value}
}
\value{
Completed dataframe with lu.from, times columns and all combinations

Internal function. Adds missing columns and completes potential sparse dataframes.
}
\description{
Complete input targets
}
\keyword{internal}
