% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_input.R
\name{complete_targets_pop}
\alias{complete_targets_pop}
\title{Complete input targets for population downscaling}
\usage{
complete_targets_pop(targets)
}
\arguments{
\item{targets}{Dataframe of targets, must have columns pop.type and value}
}
\value{
Completed dataframe with pop.type, times columns and all combinations

Internal function. Adds missing columns and completes potential sparse dataframes.
}
\description{
Complete input targets for population downscaling
}
\keyword{internal}
