% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_input.R
\name{complete_xmat.coltypes}
\alias{complete_xmat.coltypes}
\title{Complete input xmat.coltypes}
\usage{
complete_xmat.coltypes(xmat.coltypes, xmat)
}
\arguments{
\item{xmat.coltypes}{Dataframe of xmat.coltypes, must have columns ks and value;
if it is NULL, this will be created as all static
deprecated: can also be a vector}
}
\value{
Completed dataframe with ks and values

Internal function. Adds missing columns and completes potential sparse dataframes.
}
\description{
Complete input xmat.coltypes
}
\keyword{internal}
