% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_input.R
\name{complete_xmat.proj}
\alias{complete_xmat.proj}
\title{Complete input xmat.proj}
\usage{
complete_xmat.proj(xmat.proj)
}
\arguments{
\item{xmat.proj}{Dataframe of xmat.proj, must have columns ns, ks and value}
}
\value{
Completed dataframe with times and all combinations

Internal function. Adds missing columns and completes potential sparse dataframes.
}
\description{
Complete input xmat.proj
}
\keyword{internal}
