% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_input.R
\name{complete_xmat}
\alias{complete_xmat}
\title{Complete input xmat}
\usage{
complete_xmat(xmat)
}
\arguments{
\item{xmat}{Dataframe of xmat, must have columns ns, ks and value}
}
\value{
Completed dataframe

Internal function. Placeholder in case of needed additional completions.
}
\description{
Complete input xmat
}
\keyword{internal}
