% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale_control.R
\name{downscale_control}
\alias{downscale_control}
\title{Set options for downscaling}
\usage{
downscale_control(
  solve_fun = "solve_biascorr",
  algorithm = "NLOPT_LN_SBPLX",
  xtol_rel = 1e-20,
  xtol_abs = 1e-20,
  maxeval = 1600,
  MAX_EXP = log(.Machine$double.xmax),
  cutoff = 0,
  redo = 2,
  max_diff = 1e-08,
  ref_class_adjust_threshold = 1e-08,
  err.txt = ""
)
}
\arguments{
\item{solve_fun}{The name of the downscaling function to use. Has to be a
valid R function. Defaults to \code{"solve_biascorr"}.}

\item{algorithm}{Solver algorithm (see the \code{\link[nloptr]{nloptr}} package for documentation and more detail).}

\item{xtol_rel}{Relative tolerance of solver.}

\item{xtol_abs}{Absolute solver tolerance.}

\item{maxeval}{Maximum evaluation of solver.}

\item{MAX_EXP}{Numerical cutoff for MNL function.}

\item{cutoff}{Optional cutoff to avoid MNL values close to zero.}

\item{redo}{Maximum number of repeats}

\item{max_diff}{If difference to targets is larger, redo the estimation (helps to avoid convergence errors)}

\item{ref_class_adjust_threshold}{IF the share of the implied target for the multinomial logit
reference class is smaller than this threshold, assitional areas will be artificially created
(and automatically removed after solve) for numerical stability. Default to \code{1.0e-8}.}

\item{err.txt}{Error text for caller identification (used for debugging)}
}
\value{
List with default options for bias correction solver
}
\description{
This function can be called to specify additional options to the downscalR solver.
The user can change the solver function, specify solver behavior, and stopping conditions.
}
\details{
A user specified solve_fun function has to be able to process the following input parameters:
\code{targets}, \code{areas}, \code{xmat}, \code{betas}, \code{priors}, \code{restrictions}, and
\code{options} (the output of this function). The output should be a list containing two parameters:
a dataframe \code{out.res} containing the downscaled output and a list \code{out.solver} detailing
the solver results.
}
\examples{
opts1 = downscale_control()
}
