% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale_control.R
\name{downscale_control_pop}
\alias{downscale_control_pop}
\title{Set options for population downscaling}
\usage{
downscale_control_pop(
  solve_fun = "solve_biascorr",
  MAX_EXP = log(.Machine$double.xmax),
  max_diff = 1e-08,
  err.txt = ""
)
}
\arguments{
\item{solve_fun}{The name of the downscaling function to use. Has to be a
valid R function. Defaults to \code{"solve_biascorr"}.}

\item{MAX_EXP}{Numerical cutoff for MNL function.}

\item{max_diff}{If difference to targets is larger, redo the estimation (helps to avoid convergence errors)}

\item{err.txt}{Error text for caller identification (used for debugging)}
}
\value{
List with default options for bias correction solver
}
\description{
This function can be called to specify additional options to the downscalR solver.
The user can change the solver function, specify solver behavior, and stopping conditions.
}
\examples{
opts1 = downscale_control_pop()
}
