% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err_check_input.R
\name{err_check_inputs_pop}
\alias{err_check_inputs_pop}
\title{Error check inputs for population downscaling}
\usage{
err_check_inputs_pop(
  targets,
  xmat,
  betas,
  xmat.coltypes,
  xmat.proj,
  xmat.dyn.fun,
  err.txt
)
}
\arguments{
\item{targets}{A dataframe with columns times, lu.from (optional), lu.to and value (all targets >= 0)}

\item{xmat}{A dataframe of explanatory variables with columns ks and value}

\item{betas}{A dataframe of coefficients with columns ks, lu.from (optional), lu.to & value}

\item{xmat.coltypes}{A dataframce with columns ks and string value, can be either "static", "dynamic", or "projected"}

\item{xmat.proj}{dataframe with columns times, ns, ks, must be present for each xmat.coltype specified as projected}

\item{xmat.dyn.fun}{function providing update for dynamic xmat columns, must take as arguments res, curr.areas, priors, xmat.proj must return ns x ks(dynamic) columns}

\item{priors}{A dataframe of priors (if no \code{betas} were supplied) with columns ns, lu.from (optional), lu.to (with priors >= 0)}

\item{restrictions}{A dataframe with columns ns, lu.from (optional), lu.to and value. Values must be zero or one. If restrictions are one, the MNL function is set to zero

Internal function. Must throw errors, no return value if inputs do not match the specification. Handle all error checking here
Use this for all error checking of inputs.}
}
\description{
Error check inputs for population downscaling
}
\keyword{internal}
