% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fienberg.R
\name{fienberg}
\alias{fienberg}
\title{Fienberg rebalancing for two vectors with a given starting matrix}
\usage{
fienberg(start_mat, target_from, target_to)
}
\arguments{
\item{start_mat}{An \eqn{n} by \eqn{n} matrix with non-negative elements
and row sums equal to 1.}

\item{target_from}{An \eqn{n} by 1 vector with values >=0.}

\item{target_to}{An \eqn{n} by 1 vector with values >=0. The sum of
\code{target_from} must be equal to the sum of \code{target_to}.

Implements a specific version of the iterative proportional fitting algorithm.
Given an \eqn{n} by \eqn{n} matrix \code{start_mat}, the algorithm rebalances it
so that the \code{target_from * start_mat = target_to}. This is done by an iterative
fitting algorithm.}
}
\value{
The rebalanced \eqn{n} by \eqn{n} matrix \code{start_mat}.
}
\description{
Fienberg rebalancing for two vectors with a given starting matrix
}
\examples{
set.seed(123)
n = 10
vec0 = runif(n) * 10
vec1 = runif(n); vec1 = vec1 / sum(vec1) * sum(vec0)
matA = diag(n) + 10^-8; matA = matA / rowSums(matA)

res = fienberg(start_mat=matA,target_from=vec0,target_to=vec1)
}
