% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_biascorr.R
\name{get_opts.solve_biascorr}
\alias{get_opts.solve_biascorr}
\title{Get default options for bias corrections solver}
\usage{
get_opts.solve_biascorr()
}
\value{
List with default options for bias correction solver
}
\description{
Get default options for bias corrections solver
}
\details{
Call this function if you want to change default options for the bias corrections solver.

The default algorithm is NLOPT_LB_SBPLX (see the \code{\link[nloptr]{nloptr}} package for documentation and more detail). Relative and absolute tolerance are set to 1.0e-20 and \code{maxeval} to 1600.
}
\examples{
opts1 = get_opts.solve_biascorr()
}
