% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu.mnl.R
\name{grad_sqr_diff.mnl}
\alias{grad_sqr_diff.mnl}
\title{Provides analytical gradients for the squared differences optimization function with multinomial logit}
\usage{
grad_sqr_diff.mnl(x, mu, areas, targets, restrictions = NULL, cutoff = 0)
}
\arguments{
\item{x}{Bias correction parameters}

\item{mu}{Means of each class}

\item{areas}{Vector of areas}

\item{targets}{Targets to match}

\item{restrictions}{Optional restrictions. Binary, if 1, the log-odds are set to zero}

\item{cutoff}{Optional, defaults to zero. If set, all log-odds below cutoff are set to zero}
}
\value{
Vector of length \eqn{J} containing the gradients for each value of \code{x}.

Does not work for \code{cutoff} \eqn{> 0} or with restrictions. Generates a warning if these are supplied.
}
\description{
Provides analytical gradients for the squared differences optimization function with multinomial logit
}
\keyword{internal}
