% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LUC_plot.R
\name{LUC_plot}
\alias{LUC_plot}
\title{Plot function to visualize downscale results}
\usage{
LUC_plot(
  res,
  rasterfile,
  year = NULL,
  LU = NULL,
  color = "Greens",
  label = "Area in ha per pixel"
)
}
\arguments{
\item{res}{Result from downscale}

\item{rasterfile}{RasterLayer object}

\item{year}{Specify dates of results to plot (has to be in res)}

\item{LU}{Specify land-use of results to plot (has to be in res)}

\item{color}{Pixel color on the map}

\item{label}{Label of plot legend}
}
\value{
A list containing
\itemize{
\item \code{LUC.plot} A ggplot object
\item \code{plot.df} A dataframe used for constructing LUC.plot
}
}
\description{
Plot function to visualize downscale results
}
\examples{
## A basic example
}
