% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu.mnl.R
\name{mu.mnl}
\alias{mu.mnl}
\title{Multinomial logit mean calculation with restrictions and cutoff}
\usage{
mu.mnl(x, mu, areas, restrictions = NULL, cutoff = 0)
}
\arguments{
\item{x}{Bias correction parameters}

\item{mu}{Means of each class}

\item{areas}{Vector of areas}

\item{restrictions}{Optional restrictions. Binary, if 1, the log-odds are set to zero}

\item{cutoff}{Optional, defaults to zero. If set, all log-odds below cutoff are set to zero}
}
\value{
Returns the vector of log-odds times the areas
}
\description{
Multinomial logit mean calculation with restrictions and cutoff
}
\keyword{internal}
