% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_notarget.R
\name{solve_notarget.mnl}
\alias{solve_notarget.mnl}
\title{Solver for multinomial logit type problems, using only prior module projections}
\usage{
solve_notarget.mnl(
  targets,
  areas,
  xmat,
  betas,
  restrictions = NULL,
  options = downscale_control()
)
}
\arguments{
\item{targets}{A dataframe with columns lu.from, lu.to and value (all targets will be ignored)}

\item{areas}{A dataframe of areas with columns lu.from, ns and value, with all areas >= 0
and with sum(areas) >= sum(targets)}

\item{xmat}{A dataframe of explanatory variables with columns ks and value.}

\item{betas}{A dataframe of coefficients with columns ks, lu.from, lu.to & value}

\item{restrictions}{A dataframe with columns ns, lu.from, lu.to and value. Values must be zero or one. If restrictions are one, the MNL function is set to zero}

\item{options}{A list with solver options. Call \code{\link{downscale_control}} for default options and for more detail.}
}
\value{
A list containing
\itemize{
\item \code{out.res} A \code{n x p} matrix of area allocations
\item \code{out.solver} \code{NULL}, returned for compatibility
}
}
\description{
Solver for multinomial logit type problems, using only prior module projections
}
\details{
Given \code{p} targets matches the projections from an MNL-type model.

You should not call this functions directly, call \code{\link{downscale}} instead.

Areas correspond to either an areas per pixel (ns), with value or optionally the are of lu.from in a pixel. All areas must be larger The function expects lu.from
}
\examples{
## A basic example
}
