% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_input.R
\name{target_area_check}
\alias{target_area_check}
\title{Check for targets to area mismatch over timesteps}
\usage{
target_area_check(targets, areas)
}
\arguments{
\item{targets}{Dataframe of targets, must have columns lu.to and value}

\item{areas}{Dataframe of areas, must have columns ns and value#' @

Internal function. Checks if targets are fullfillable over time before downscaling.}
}
\description{
Check for targets to area mismatch over timesteps
}
\keyword{internal}
