% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_netcdf.R
\name{write_netcdf}
\alias{write_netcdf}
\title{Function to write results from downscalr into NetCDF file format}
\usage{
write_netcdf(
  res,
  rasterfile,
  filename = "nc_v1.nc",
  raster.crs = "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0",
  times = NULL,
  ns = NULL,
  lu.to = NULL,
  value = NULL,
  label = NULL,
  . = NULL
)
}
\arguments{
\item{res}{Result from downscale}

\item{rasterfile}{RasterLayer object with ns as values}

\item{filename}{Name and path of NetCDF file}

\item{raster.crs}{CRS of raster}
}
\value{
\itemize{
\item \file{write_netcdf} with name and path specified in filename function argument
}
}
\description{
Function to write results from downscalr into NetCDF file format
}
\examples{
## A basic example to plot the observed LU changes 

require(dplyr)
require(tidyr)
require(tibble)
require(ncdf4)
require(raster)

areas <- data.frame(ns=getValues(argentina_raster), 
                    area=getValues(area(argentina_raster))) \%>\%
         na.omit() \%>\% 
         group_by(ns) \%>\% 
         summarise(area=sum(area)) \%>\%
         mutate(ns=as.character(ns))
 
to.plot <- argentina_luc \%>\% 
           rename("times"="Ts") \%>\% 
           left_join(areas) \%>\% 
           mutate(value=value*area) \%>\% 
           dplyr::select(-area)
 
write_netcdf(to.plot ,argentina_raster, filename="H:/argentina.nc")

}
