% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmat_update.R
\name{xmat.sum_to}
\alias{xmat.sum_to}
\title{xmat summing up function over land-use changes}
\usage{
xmat.sum_to(res, curr.areas, priors, xmat, xmat.proj)
}
\arguments{
\item{res}{Result from downscale}

\item{curr.areas}{Dataframe of current areas}

\item{priors}{Priors for updating}

\item{xmat}{X-matrix dataframe}

\item{xmat.proj}{Projected x matrix}
}
\value{
curr.xmat Dynamically updated xmat columns
}
\description{
xmat summing up function over land-use changes
}
