% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{downscale}
\alias{downscale}
\title{Downscaling of land-use (change) data}
\usage{
downscale(
  targets,
  start.areas,
  xmat,
  betas,
  areas.update.fun = areas.sum_to,
  xmat.coltypes = NULL,
  xmat.proj = NULL,
  xmat.dyn.fun = NULL,
  priors = NULL,
  restrictions = NULL,
  options = downscale_control()
)
}
\arguments{
\item{targets}{A dataframe with columns times, lu.from (optional), lu.to and value (all targets >= 0)}

\item{start.areas}{A dataframe of areas with columns lu.from (optional), ns and value, with all areas >= 0 and with sum(areas) >= sum(targets)}

\item{xmat}{A dataframe of explanatory variables with columns ks and value}

\item{betas}{A dataframe of coefficients with columns ks, lu.from (optional), lu.to & value}

\item{areas.update.fun}{function providing update for dynamic xmat columns, must take as arguments res, curr.areas, priors, xmat.proj, must dataframe with columns ns, lu.from & value defaults to areas.sum_to() which sums over lu.to}

\item{xmat.coltypes}{ks vector, each can be either "static", "dynamic", or "projected"}

\item{xmat.proj}{dataframe with columns times, ns, ks, must be present for each xmat.coltype specified as projected}

\item{xmat.dyn.fun}{function providing update for dynamic xmat columns, must take as arguments res, curr.areas, priors, xmat.proj must return ns x ks(dynamic) columns}

\item{priors}{A dataframe of priors (if no \code{betas} were supplied) with columns ns, lu.from (optional), lu.to (with priors >= 0)}

\item{restrictions}{A dataframe with columns ns, lu.from (optional), lu.to and value. Values must be zero or one. If restrictions are one, the MNL function is set to zero}

\item{options}{A list with solver options. Call \code{\link{downscale_control}} for default options and for more detail.}
}
\value{
A list containing
\itemize{
\item \code{out.res} Dataframe with columns times, ns, lu.from, lu.to & value (area allocation)
\item \code{out.solver} A list of the solver output
}
}
\description{
Downscaling of land-use (change) data
}
\details{
Given \code{p} targets matches either the projections from an MNL-type model or exogenous priors.
}
\examples{
## A basic example
}
