% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale_control.R
\name{downscale_control}
\alias{downscale_control}
\title{Set Options for Downscaling Solver}
\usage{
downscale_control(
  solve_fun = "solve_biascorr",
  algorithm = "NLOPT_LN_SBPLX",
  xtol_rel = 1e-20,
  xtol_abs = 1e-20,
  maxeval = 1600,
  MAX_EXP = log(.Machine$double.xmax),
  cutoff = 0,
  redo = 2,
  max_diff = 1e-08,
  ref_class_adjust_threshold = 1e-08,
  err.txt = ""
)
}
\arguments{
\item{solve_fun}{The downscaling function to use. Valid options are "\code{solve_biascorr}" for bias correction and "\code{solve_notarget}" for non-targeted downscaling. Defaults to "\code{solve_biascorr}".}

\item{algorithm}{Specifies the solver algorithm. The default algorithm is "\code{NLOPT_LN_SBPLX}". For more information on available algorithms, refer to the \code{\link[nloptr]{nloptr}} package.}

\item{xtol_rel}{The relative tolerance level of the solver, controlling solution precision in relative terms. Default is \code{1.0e-20}, where a lower value indicates higher precision.}

\item{xtol_abs}{The absolute tolerance level of the solver, specifying solution precision in absolute terms. Default is \code{1.0e-20}, with a lower value indicating higher precision.}

\item{maxeval}{The maximum number of evaluations the solver will perform, limiting computational effort. Default is \code{1600}.}

\item{MAX_EXP}{Sets the numerical cutoff to prevent extremely high values in the Multinomial Logit (MNL) function. Default is \code{log(.Machine$double.xmax)}.}

\item{cutoff}{An optional cutoff in the share of a pixel to avoid MNL values close to zero. Default is \code{0}.}

\item{redo}{The maximum number of estimation repeats to correct for potential convergence errors. Default is \code{2}.}

\item{max_diff}{The threshold for acceptable difference to targets. If exceeded, the estimation is redone. Default is \code{1.0e-8}.}

\item{ref_class_adjust_threshold}{The threshold for adjusting the reference class in the MNL. If the implied target share for the reference class is below this, additional areas are created for numerical stability and removed after solving. Default is \code{1.0e-8}.}

\item{err.txt}{An optional error message for debugging purposes. Empty by default.}
}
\value{
Returns a list with the specified options for the downscaling solver.
}
\description{
This function configures the downscalR solver with customizable options. It allows the selection of the solver function and adjustment of solver behavior and stopping conditions, tailored for specific downscaling needs.
}
\details{
The function returns a list of parameters to be used by the solver. It is essential to provide compatible parameters to ensure accurate downscaling results. The function will stop with an error message if \code{solve_fun} is not correctly specified.
}
\examples{
# Default settings
opts_default = downscale_control()

# Custom settings
opts_custom = downscale_control(algorithm = "some_other_algorithm",
                                xtol_rel = 1e-5, xtol_abs = 1e-5,
                                maxeval = 2000, redo = 3)
}
